Detect stuff and Edit mode
I used a button instead of tap gesture recognizer because it never was
consistent the way it fired vs triggering actionable data. Also, once
in edit mode it was nearly impossible to get out without a button
anyway. I tried double taps, and multi touch taps, but at least in the
simulator it wasn’t always accurate, sometimes it wanted to just
highlight text instead of triggering the gesture recognizer.